requirements.txt
pyyaml
rapidfuzz
unidecode
sentence-transformers
openai